server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bidict==0.22.1
cffi==1.16.0
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.37.0
typing_extensions==4.9.0
uvicorn==0.27.0.post1
websockets==12.0
wsproto==1.2.0
